chore(lint): 首次接入 clippy 并清理全部 54 条告警
此前工具链未安装 clippy 组件，仓库从未跑过该门禁，因此这批告警并非本次改动引入，而是既有代码一直未度量的问题。全量为机械型 lint，无语义变更：同类型 cast（f32/usize 冗余 as）、冗余闭包（ok_or_else/map_err 直接传构造器）、len()==0 与 len()>0 改 is_empty、chars().all(is_ascii) 改 str::is_ascii、Copy 类型多余 clone、测试里 assert_eq!(x, true/false) 改 assert!、.err().expect() 改 expect_err、以及一处 !(a > b) 反比较改 a <= b。

其中 53 处由 rustfix 自动应用（cargo clippy --fix --all-targets），1 处（tests/edit_ext.rs 零宽矩形断言）因涉及偏序语义差异人工改写。

验证：cargo clippy --all-targets 零告警；cargo test -- --test-threads=1 仍 343 passed / 0 failed。

+++ src-tauri/src/edit_ext.rs
@@ -136,10 +136,10 @@ pub fn rewrite_text_logic(
                 }
                 let Ok(bounds) = obj.bounds() else { continue };
                 let (l, b, r, t) = (
-                    bounds.left().value as f32,
-                    bounds.bottom().value as f32,
-                    bounds.right().value as f32,
-                    bounds.top().value as f32,
+                    bounds.left().value,
+                    bounds.bottom().value,
+                    bounds.right().value,
+                    bounds.top().value,
                 );
                 let obj_rect = PtRect {
                     left: l.min(r),
@@ -317,10 +317,10 @@ pub async fn replace_image(
             objs.get(object_index as usize).ok().and_then(|obj| {
                 obj.bounds().ok().map(|b| {
                     (
-                        b.left().value as f32,
-                        b.bottom().value as f32,
-                        b.right().value as f32,
-                        b.top().value as f32,
+                        b.left().value,
+                        b.bottom().value,
+                        b.right().value,
+                        b.top().value,
                     )
                 })
             })
@@ -458,10 +458,10 @@ pub fn list_image_objects_logic(
         }
         let Ok(b) = obj.bounds() else { continue };
         let (l, bb, r, t) = normalize_bounds(
-            b.left().value as f32,
-            b.bottom().value as f32,
-            b.right().value as f32,
-            b.top().value as f32,
+            b.left().value,
+            b.bottom().value,
+            b.right().value,
+            b.top().value,
         );
         out.push(ImageObjectInfo {
             object_index: i as u32,
